completed side-bar
