+++ app/src/components/side-bar/sidebar-item/styles.css
@@ -1,0 +1,76 @@
+.side-bar-item {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-end;
+  height: 80px;
+  width: 100%;
+}
+
+.side-bar-item>a {
+  flex-shrink: 0;
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  align-items: center;
+  background-color: transparent;
+  height: 40px;
+  width: 90%;
+  border-radius: 40px 0 0 40px;
+  text-decoration: none;
+  font-weight: 500;
+  font-size: 1rem;
+  color: white;
+}
+
+@media screen and (max-width: 1350px) {
+  .side-bar-item>a {
+    font-size: 0.8rem;
+  }
+}
+
+.side-bar-item.selected>a {
+  position: relative;
+  background-color: var(--selected-side-bar);
+  font-weight: 600;
+  color: var(--selected-side-bar-text);
+}
+
+.side-bar-item.selected>a::before {
+  content: "";
+  position: absolute;
+  background-color: transparent;
+  top: -40px;
+  right: 0;
+  height: 40px;
+  width: 25px;
+  border-bottom-right-radius: 25px;
+  box-shadow: 0 25px 0 0 var(--selected-side-bar);
+}
+
+.side-bar-item.selected>a::after {
+  content: "";
+  position: absolute;
+  background-color: transparent;
+  bottom: -40px;
+  right: 0;
+  height: 40px;
+  width: 25px;
+  border-top-right-radius: 25px;
+  box-shadow: 0 -25px 0 0 var(--selected-side-bar);
+}
+
+.side-bar-item>a>* {
+  margin-left: 0.5vw;
+  z-index: 1;
+}
+
+.side-bar-item>a>*:first-child {
+  margin-left: 1vw;
+}
+
+.side-bar-item>a>svg {
+  flex-shrink: 0;
+  width: 15px;
+  height: auto;
+}
