+++ app/src/pages/Dashboard/styles.css
@@ -1,0 +1,23 @@
+.dashboard {
+  background-image: url(../../assets/images/background.png), linear-gradient(to bottom right, var(--background-color-1), var(--background-color-2));
+  min-height: 100vh;
+}
+
+.dashboard-container {
+  padding-top: 32px;
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+}
+
+.dashboard-body {
+  width: 100%;
+  border: 1px solid red;
+}
+
+.dashboard-header {
+  height: 88px;
+  width: 100%;
+  background-color: white;
+  border-radius: 30px 30px 0 0;
+}
