--- app/src/pages/Home/styles.css
@@ -1,56 +1,0 @@
-.home {
-  background-image: url(../../assets/images/background.png), linear-gradient(to bottom right, var(--background-color-1), var(--background-color-2));
-  min-height: 100vh;
-}
-
-.home-container {
-  padding-top: 32px;
-  width: 100%;
-  border: 1px solid red;
-  display: flex;
-  flex-direction: row;
-}
-
-.side-bar {
-  display: flex;
-  flex-direction: column;
-  border: 2px solid lime;
-  width: 15vw;
-}
-
-.side-bar-item {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  height: 80px;
-  width: 100%;
-}
-
-.side-bar-item>a {
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-  background-color: transparent;
-  height: 40px;
-  width: 90%;
-  border-radius: 40px 0 0 40px;
-  text-decoration: none;
-  font-weight: 500;
-  font-size: 1rem;
-  color: white;
-}
-
-.side-bar-item.selected>a {
-  background-color: var(--selected-side-bar);
-  font-weight: 600;
-  color: var(--selected-side-bar-text);
-}
-
-.side-bar-item>a>* {
-  margin-left: 16px;
-}
-
-.side-bar-item>a>*:first-child {
-  margin-left: 24px;
-}
