+++ app/src/routes.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 
-import Home from './pages/Home';
+import Dashboard from './pages/Dashboard';
 
 const DefaultRoutes = () => (
   <BrowserRouter>
     <Switch>
-      <Route exact path="/" component={Home} />
+      <Route exact path="/" component={Dashboard} />
     </Switch>
   </BrowserRouter>
 );
